requirements.txt
python-dotenv
openai
google-cloud-vision
